fix: fixed files name inside tests/ and inserted token_test inside tests/
