+++ src/tests/mod.rs
@@ -1,3 +1,4 @@
-mod compiler_error;
-mod lexer_test;
-mod source_error;
+mod compiler_test;
+mod lexical_test;
+mod source_test;
+mod token_test;
